+++ server/src/main/java/com/example/clinisched/services/patientService/PatientService.java
@@ -1,8 +1,10 @@
 package com.example.clinisched.services.patientService;
 
-import com.example.clinisched.exception.NotFoundException;
+import com.example.clinisched.dto.patientDTO.PatientRequestDTO;
+import com.example.clinisched.dto.patientDTO.PatientResponseDTO;
 import com.example.clinisched.models.patientModel.PatientModel;
 import com.example.clinisched.repositories.patientRepository.PatientRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,16 +21,37 @@ public class PatientService {
         return patientRepository.findAll();
     }
 
-    public PatientModel getById(Long id){
-        return patientRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException(id));
+    public PatientResponseDTO getById(Long id){
+        PatientModel patientModel = patientRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Patient with ID " + id + " not found!"));
+        return new PatientResponseDTO(patientModel);
     }
 
-    public PatientModel post(PatientModel patientModel){
-        return patientRepository.save(patientModel);
+    public PatientModel post(PatientRequestDTO patientRequestDTO){
+        try{
+            PatientModel patientModel = new PatientModel(patientRequestDTO);
+            return patientRepository.save(patientModel);
+        } catch (Exception e) {
+            throw new RuntimeException("Error to save patient. Try again later!");
+        }
     }
+
+    public PatientModel update(Long id, PatientRequestDTO patientRequestDTO) {
+        PatientModel existingPatient = patientRepository.findById(patientRequestDTO.getId())
+                .orElseThrow(() -> new EntityNotFoundException("Patient with ID " + id + " not found!"));
+
+        existingPatient.setName(patientRequestDTO.getName());
+        existingPatient.setEmail(patientRequestDTO.getEmail());
+        existingPatient.setPhone(patientRequestDTO.getPhone());
+
+        return patientRepository.save(existingPatient);
+    }
+
 
     public void delete(Long id){
+        if(!patientRepository.existsById(id)){
+            throw new EntityNotFoundException("Patient with ID " + id + " not found!");
+        }
         patientRepository.deleteById(id);
     }
